requirements.txt
requests==2.22.0
pyyaml==5.1.1
